refactor vasu document and template

+++ service/src/main/kotlin/fi/espoo/evaka/vasu/VasuTemplateController.kt
@@ -5,6 +5,7 @@
 package fi.espoo.evaka.vasu
 
 import fi.espoo.evaka.Audit
+import fi.espoo.evaka.shared.VasuTemplateId
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
 import fi.espoo.evaka.shared.auth.UserRole
 import fi.espoo.evaka.shared.db.Database
@@ -22,7 +23,6 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
-import java.util.UUID
 
 @RestController
 @RequestMapping("/vasu/templates")
@@ -40,7 +40,7 @@ class VasuTemplateController(
         db: Database.Connection,
         user: AuthenticatedUser,
         @RequestBody body: CreateTemplateRequest
-    ): UUID {
+    ): VasuTemplateId {
         Audit.VasuTemplateCreate.log()
         accessControl.requirePermissionFor(user, Action.Global.CREATE_VASU_TEMPLATE)
 
@@ -58,11 +58,11 @@ class VasuTemplateController(
     fun editTemplate(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID,
+        @PathVariable id: VasuTemplateId,
         @RequestBody body: VasuTemplateUpdate
     ) {
         Audit.VasuTemplateEdit.log()
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.UPDATE, id)
 
         db.transaction { tx ->
             val template = tx.getVasuTemplateForUpdate(id) ?: throw NotFound("Template not found")
@@ -80,11 +80,11 @@ class VasuTemplateController(
     fun copyTemplate(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID,
+        @PathVariable id: VasuTemplateId,
         @RequestBody body: CopyTemplateRequest
-    ): UUID {
+    ): VasuTemplateId {
         Audit.VasuTemplateCopy.log(id)
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.COPY, id)
 
         return db.transaction {
             val template = it.getVasuTemplate(id) ?: throw NotFound("template not found")
@@ -104,12 +104,7 @@ class VasuTemplateController(
         @RequestParam(required = false) validOnly: Boolean = false
     ): List<VasuTemplateSummary> {
         Audit.VasuTemplateRead.log()
-        user.requireOneOfRoles(
-            UserRole.ADMIN,
-            UserRole.UNIT_SUPERVISOR,
-            UserRole.SPECIAL_EDUCATION_TEACHER,
-            UserRole.STAFF
-        )
+        accessControl.requirePermissionFor(user, Action.Global.READ_VASU_TEMPLATE)
 
         return db.read { tx -> tx.getVasuTemplates(validOnly) }
     }
@@ -118,10 +113,11 @@ class VasuTemplateController(
     fun getTemplate(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID
+        @PathVariable id: VasuTemplateId
     ): VasuTemplate {
         Audit.VasuTemplateRead.log(id)
         user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.READ, id)
 
         return db.read { tx -> tx.getVasuTemplate(id) } ?: throw NotFound("template $id not found")
     }
@@ -130,10 +126,10 @@ class VasuTemplateController(
     fun deleteTemplate(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID
+        @PathVariable id: VasuTemplateId
     ) {
         Audit.VasuTemplateDelete.log(id)
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.DELETE, id)
 
         db.transaction { it.deleteUnusedVasuTemplate(id) }
     }
@@ -142,11 +138,11 @@ class VasuTemplateController(
     fun putTemplateContent(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID,
+        @PathVariable id: VasuTemplateId,
         @RequestBody content: VasuContent
     ) {
         Audit.VasuTemplateUpdate.log(id)
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.UPDATE, id)
 
         db.transaction { tx ->
             val template = tx.getVasuTemplateForUpdate(id) ?: throw NotFound("template $id not found")
